+++ jsonapi-resources.gemspec
@@ -28,6 +28,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency 'pry-byebug'
   spec.add_development_dependency 'concurrent-ruby-ext'
   spec.add_development_dependency 'sequel'
+  spec.add_development_dependency 'sequel-rails'
   spec.add_development_dependency 'activerecord', '>= 4.1'
   spec.add_dependency 'activesupport', '>= 4.1'
   spec.add_dependency 'railties', '>= 4.1'
