+++ test/support/controllers_resources_processors.rb
@@ -1,659 +1,4 @@
-require 'active_record'
-require 'jsonapi-resources'
-
-ActiveSupport::Inflector.inflections(:en) do |inflect|
-  inflect.uncountable 'preferences'
-  inflect.irregular 'numero_telefone', 'numeros_telefone'
-end
-
-### DATABASE
-ActiveRecord::Schema.define do
-  create_table :people, force: true do |t|
-    t.string     :name
-    t.string     :email
-    t.datetime   :date_joined
-    t.belongs_to :preferences
-    t.integer    :hair_cut_id, index: true
-    t.boolean    :book_admin, default: false
-    t.boolean    :special, default: false
-    t.timestamps null: false
-  end
-
-  create_table :author_details, force: true do |t|
-    t.integer :person_id
-    t.string  :author_stuff
-  end
-
-  create_table :posts, force: true do |t|
-    t.string     :title, length: 255
-    t.text       :body
-    t.integer    :author_id
-    t.integer    :parent_post_id
-    t.belongs_to :section, index: true
-    t.timestamps null: false
-  end
-
-  create_table :comments, force: true do |t|
-    t.text       :body
-    t.belongs_to :post, index: true
-    t.integer    :author_id
-    t.timestamps null: false
-  end
-
-  create_table :companies, force: true do |t|
-    t.string     :type
-    t.string     :name
-    t.string     :address
-    t.timestamps null: false
-  end
-
-  create_table :tags, force: true do |t|
-    t.string :name
-    t.timestamps null: false
-  end
-
-  create_table :sections, force: true do |t|
-    t.string :name
-    t.timestamps null: false
-  end
-
-  create_table :posts_tags, force: true do |t|
-    t.references :post, :tag, index: true
-  end
-  add_index :posts_tags, [:post_id, :tag_id], unique: true
-
-  create_table :special_post_tags, force: true do |t|
-    t.references :post, :tag, index: true
-  end
-  add_index :special_post_tags, [:post_id, :tag_id], unique: true
-
-  create_table :comments_tags, force: true do |t|
-    t.references :comment, :tag, index: true
-  end
-
-  create_table :iso_currencies, id: false, force: true do |t|
-    t.string :code, limit: 3, null: false
-    t.string :name
-    t.string :country_name
-    t.string :minor_unit
-    t.timestamps null: false
-  end
-  add_index :iso_currencies, :code, unique: true
-
-  create_table :expense_entries, force: true do |t|
-    t.string :currency_code, limit: 3, null: false
-    t.integer :employee_id, null: false
-    t.decimal :cost, precision: 12, scale: 4, null: false
-    t.date :transaction_date
-    t.timestamps null: false
-  end
-
-  create_table :planets, force: true do |t|
-    t.string :name
-    t.string :description
-    t.integer :planet_type_id
-  end
-
-  create_table :planets_tags, force: true do |t|
-    t.references :planet, :tag, index: true
-  end
-  add_index :planets_tags, [:planet_id, :tag_id], unique: true
-
-  create_table :planet_types, force: true do |t|
-    t.string :name
-  end
-
-  create_table :moons, force: true do |t|
-    t.string  :name
-    t.string  :description
-    t.integer :planet_id
-    t.timestamps null: false
-  end
-
-  create_table :craters, id: false, force: true do |t|
-    t.string  :code
-    t.string  :description
-    t.integer :moon_id
-    t.timestamps null: false
-  end
-
-  create_table :preferences, force: true do |t|
-    t.integer :person_id
-    t.boolean :advanced_mode, default: false
-    t.timestamps null: false
-  end
-
-  create_table :facts, force: true do |t|
-    t.integer  :person_id
-    t.string   :spouse_name
-    t.text     :bio
-    t.float    :quality_rating
-    t.decimal  :salary, precision: 12, scale: 2
-    t.datetime :date_time_joined
-    t.date     :birthday
-    t.time     :bedtime
-    t.binary   :photo, limit: 1.kilobyte
-    t.boolean  :cool
-    t.timestamps null: false
-  end
-
-  create_table :books, force: true do |t|
-    t.string :title
-    t.string :isbn
-    t.boolean :banned, default: false
-    t.timestamps null: false
-  end
-
-  create_table :book_authors, force: true do |t|
-    t.integer :book_id
-    t.integer :person_id
-  end
-
-  create_table :book_comments, force: true do |t|
-    t.text       :body
-    t.belongs_to :book, index: true
-    t.integer    :author_id
-    t.boolean    :approved, default: true
-    t.timestamps null: false
-  end
-
-  create_table :customers, force: true do |t|
-    t.string   :name
-    t.timestamps null: false
-  end
-
-  create_table :purchase_orders, force: true do |t|
-    t.date     :order_date
-    t.date     :requested_delivery_date
-    t.date     :delivery_date
-    t.integer  :customer_id
-    t.string   :delivery_name
-    t.string   :delivery_address_1
-    t.string   :delivery_address_2
-    t.string   :delivery_city
-    t.string   :delivery_state
-    t.string   :delivery_postal_code
-    t.float    :delivery_fee
-    t.float    :tax
-    t.float    :total
-    t.timestamps null: false
-  end
-
-  create_table :order_flags, force: true do |t|
-    t.string :name
-  end
-
-  create_table :purchase_orders_order_flags, force: true do |t|
-    t.references :purchase_order, :order_flag, index: true
-  end
-  add_index :purchase_orders_order_flags, [:purchase_order_id, :order_flag_id], unique: true, name: "po_flags_idx"
-
-  create_table :line_items, force: true do |t|
-    t.integer  :purchase_order_id
-    t.string   :part_number
-    t.string   :quantity
-    t.float    :item_cost
-    t.timestamps null: false
-  end
-
-  create_table :hair_cuts, force: true do |t|
-    t.string :style
-  end
-
-  create_table :numeros_telefone, force: true do |t|
-    t.string   :numero_telefone
-    t.timestamps null: false
-  end
-
-  create_table :categories, force: true do |t|
-    t.string :name
-    t.string :status, limit: 10
-    t.timestamps null: false
-  end
-
-  create_table :pictures, force: true do |t|
-    t.string  :name
-    t.integer :imageable_id
-    t.string  :imageable_type
-    t.timestamps null: false
-  end
-
-  create_table :documents, force: true do |t|
-    t.string  :name
-    t.timestamps null: false
-  end
-
-  create_table :products, force: true do |t|
-    t.string  :name
-    t.timestamps null: false
-  end
-
-  create_table :vehicles, force: true do |t|
-    t.string :type
-    t.string :make
-    t.string :model
-    t.string :length_at_water_line
-    t.string :drive_layout
-    t.string :serial_number
-    t.integer :person_id
-    t.timestamps null: false
-  end
-
-  create_table :makes, force: true do |t|
-    t.string :model
-    t.timestamps null: false
-  end
-
-  # special cases - fields that look like they should be reserved names
-  create_table :hrefs, force: true do |t|
-    t.string :name
-    t.timestamps null: false
-  end
-
-  create_table :links, force: true do |t|
-    t.string :name
-    t.timestamps null: false
-  end
-
-  create_table :web_pages, force: true do |t|
-    t.string :href
-    t.string :link
-    t.timestamps null: false
-  end
-
-  create_table :questionables, force: true do |t|
-    t.timestamps null: false
-  end
-
-  create_table :boxes, force: true  do |t|
-    t.string :name
-    t.timestamps null: false
-  end
-
-  create_table :things, force: true  do |t|
-    t.string :name
-    t.references :user
-    t.references :box
-
-    t.timestamps null: false
-  end
-
-  create_table :users, force: true  do |t|
-    t.string :name
-    t.timestamps null: false
-  end
-
-  create_table :related_things, force: true  do |t|
-    t.string :name
-    t.references :from, references: :thing
-    t.references :to, references: :thing
-
-    t.timestamps null: false
-  end
-
-  create_table :questions, force: true do |t|
-    t.string :text
-  end
-
-  create_table :answers, force: true do |t|
-    t.references :question
-    t.integer :respondent_id
-    t.string  :respondent_type
-    t.string :text
-  end
-
-  create_table :patients, force: true do |t|
-    t.string :name
-  end
-
-  create_table :doctors, force: true do |t|
-    t.string :name
-  end
-
-  # special cases
-end
-
-### MODELS
-class Person < ActiveRecord::Base
-  has_many :posts, foreign_key: 'author_id'
-  has_many :comments, foreign_key: 'author_id'
-  has_many :expense_entries, foreign_key: 'employee_id', dependent: :restrict_with_exception
-  has_many :vehicles
-  belongs_to :preferences
-  belongs_to :hair_cut
-  has_one :author_detail
-
-  has_and_belongs_to_many :books, join_table: :book_authors
-
-  has_many :even_posts, -> { where('posts.id % 2 = 0') }, class_name: 'Post', foreign_key: 'author_id'
-  has_many :odd_posts, -> { where('posts.id % 2 = 1') }, class_name: 'Post', foreign_key: 'author_id'
-
-  ### Validations
-  validates :name, presence: true
-  validates :date_joined, presence: true
-end
-
-class AuthorDetail < ActiveRecord::Base
-  belongs_to :author, class_name: 'Person', foreign_key: 'person_id'
-end
-
-class Post < ActiveRecord::Base
-  belongs_to :author, class_name: 'Person', foreign_key: 'author_id'
-  belongs_to :writer, class_name: 'Person', foreign_key: 'author_id'
-  has_many :comments
-  has_and_belongs_to_many :tags, join_table: :posts_tags
-  has_many :special_post_tags, source: :tag
-  has_many :special_tags, through: :special_post_tags, source: :tag
-  belongs_to :section
-  has_one :parent_post, class_name: 'Post', foreign_key: 'parent_post_id'
-
-  validates :author, presence: true
-  validates :title, length: { maximum: 35 }
-
-  before_destroy :destroy_callback
-
-  def destroy_callback
-    if title == "can't destroy me"
-      errors.add(:title, "can't destroy me")
-
-      # :nocov:
-      if Rails::VERSION::MAJOR >= 5
-        throw(:abort)
-      else
-        return false
-      end
-      # :nocov:
-    end
-  end
-end
-
-class SpecialPostTag < ActiveRecord::Base
-  belongs_to :tag
-  belongs_to :post
-end
-
-class Comment < ActiveRecord::Base
-  belongs_to :author, class_name: 'Person', foreign_key: 'author_id'
-  belongs_to :post
-  has_and_belongs_to_many :tags, join_table: :comments_tags
-end
-
-class Company < ActiveRecord::Base
-end
-
-class Firm < Company
-end
-
-class Tag < ActiveRecord::Base
-  has_and_belongs_to_many :posts, join_table: :posts_tags
-  has_and_belongs_to_many :planets, join_table: :planets_tags
-end
-
-class Section < ActiveRecord::Base
-  has_many :posts
-end
-
-class HairCut < ActiveRecord::Base
-  has_many :people
-end
-
-class Property < ActiveRecord::Base
-end
-
-class Customer < ActiveRecord::Base
-end
-
-class BadlyNamedAttributes < ActiveRecord::Base
-end
-
-class Cat < ActiveRecord::Base
-end
-
-class IsoCurrency < ActiveRecord::Base
-  self.primary_key = :code
-  # has_many :expense_entries, foreign_key: 'currency_code'
-end
-
-class ExpenseEntry < ActiveRecord::Base
-  belongs_to :employee, class_name: 'Person', foreign_key: 'employee_id'
-  belongs_to :iso_currency, foreign_key: 'currency_code'
-end
-
-class Planet < ActiveRecord::Base
-  has_many :moons
-  belongs_to :planet_type
-
-  has_and_belongs_to_many :tags, join_table: :planets_tags
-
-  # Test model callback cancelling save
-  before_save :check_not_pluto
-
-  def check_not_pluto
-    # Pluto can't be a planet, so cancel the save
-    if name.downcase == 'pluto'
-      # :nocov:
-      if Rails::VERSION::MAJOR >= 5
-        throw(:abort)
-      else
-        return false
-      end
-      # :nocov:
-    end
-  end
-end
-
-class PlanetType < ActiveRecord::Base
-  has_many :planets
-end
-
-class Moon < ActiveRecord::Base
-  belongs_to :planet
-
-  has_many :craters
-end
-
-class Crater < ActiveRecord::Base
-  self.primary_key = :code
-
-  belongs_to :moon
-end
-
-class Preferences < ActiveRecord::Base
-  has_one :author, class_name: 'Person', :inverse_of => 'preferences'
-end
-
-class Fact < ActiveRecord::Base
-  validates :spouse_name, :bio, presence: true
-end
-
-class Like < ActiveRecord::Base
-end
-
-class Breed
-
-  def initialize(id = nil, name = nil)
-    if id.nil?
-      @id = $breed_data.new_id
-      $breed_data.add(self)
-    else
-      @id = id
-    end
-    @name = name
-    @errors = ActiveModel::Errors.new(self)
-  end
-
-  attr_accessor :id, :name
-
-  def destroy
-    $breed_data.remove(@id)
-  end
-
-  def valid?(context = nil)
-    @errors.clear
-    if name.is_a?(String) && name.length > 0
-      return true
-    else
-      @errors.add(:name, "can't be blank")
-      return false
-    end
-  end
-
-  def errors
-    @errors
-  end
-end
-
-class Book < ActiveRecord::Base
-  has_many :book_comments
-  has_many :approved_book_comments, -> { where(approved: true) }, class_name: "BookComment"
-
-  has_and_belongs_to_many :authors, join_table: :book_authors, class_name: "Person"
-end
-
-class BookComment < ActiveRecord::Base
-  belongs_to :author, class_name: 'Person', foreign_key: 'author_id'
-  belongs_to :book
-
-  def self.for_user(current_user)
-    records = self
-    # Hide the unapproved comments from people who are not book admins
-    unless current_user && current_user.book_admin
-      records = records.where(approved: true)
-    end
-    records
-  end
-end
-
-class BreedData
-  def initialize
-    @breeds = {}
-  end
-
-  def breeds
-    @breeds
-  end
-
-  def new_id
-    @breeds.keys.max + 1
-  end
-
-  def add(breed)
-    @breeds[breed.id] = breed
-  end
-
-  def remove(id)
-    @breeds.delete(id)
-  end
-end
-
-class Customer < ActiveRecord::Base
-  has_many :purchase_orders
-end
-
-class PurchaseOrder < ActiveRecord::Base
-  belongs_to :customer
-  has_many :line_items
-  has_many :admin_line_items, class_name: 'LineItem', foreign_key: 'purchase_order_id'
-
-  has_and_belongs_to_many :order_flags, join_table: :purchase_orders_order_flags
-
-  has_and_belongs_to_many :admin_order_flags, join_table: :purchase_orders_order_flags, class_name: 'OrderFlag'
-end
-
-class OrderFlag < ActiveRecord::Base
-  has_and_belongs_to_many :purchase_orders, join_table: :purchase_orders_order_flags
-end
-
-class LineItem < ActiveRecord::Base
-  belongs_to :purchase_order
-end
-
-class NumeroTelefone < ActiveRecord::Base
-end
-
-class Category < ActiveRecord::Base
-end
-
-class Picture < ActiveRecord::Base
-  belongs_to :imageable, polymorphic: true
-end
-
-class Vehicle < ActiveRecord::Base
-  belongs_to :person
-end
-
-class Car < Vehicle
-end
-
-class Boat < Vehicle
-end
-
-class Document < ActiveRecord::Base
-  has_many :pictures, as: :imageable
-end
-
-class Document::Topic < Document
-end
-
-class Product < ActiveRecord::Base
-  has_one :picture, as: :imageable
-end
-
-class Make < ActiveRecord::Base
-end
-
-class WebPage < ActiveRecord::Base
-end
-
-class Box < ActiveRecord::Base
-  has_many :things
-end
-
-class User < ActiveRecord::Base
-  has_many :things
-end
-
-class Thing < ActiveRecord::Base
-  belongs_to :box
-  belongs_to :user
-
-  has_many :related_things, foreign_key: :from_id
-  has_many :things, through: :related_things, source: :to
-end
-
-class RelatedThing < ActiveRecord::Base
-  belongs_to :from, class_name: Thing, foreign_key: :from_id
-  belongs_to :to, class_name: Thing, foreign_key: :to_id
-end
-
-class Question < ActiveRecord::Base
-  has_one :answer
-
-  def respondent
-    answer.try(:respondent)
-  end
-end
-
-class Answer < ActiveRecord::Base
-  belongs_to :question
-  belongs_to :respondent, polymorphic: true
-end
-
-class Patient < ActiveRecord::Base
-end
-
-class Doctor < ActiveRecord::Base
-end
-
-module Api
-  module V7
-    class Client < Customer
-    end
-
-    class Customer < Customer
-    end
-  end
-end
+# Controllers, Resources, and Processors for specs.
 
 ### CONTROLLERS
 class AuthorsController < JSONAPI::ResourceControllerMetal
