+++ test/support/orm/active_record/initialize.rb
@@ -1,0 +1,1 @@
+require 'active_record/railtie'
